requirements.txt
appdirs==1.4.4
black==20.8b1
certifi==2020.6.20
click==7.1.2
elasticsearch==7.9.1
epam.indigo==1.4.0b0
ftpretty==0.3.3
isort==5.5.1
mypy==0.782
mypy-extensions==0.4.3
pathspec==0.8.0
pipdeptree==1.0.0
python-dateutil==2.8.1
regex==2020.7.14
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.10
